[BE] REFACTOR: chatroom gateway 분리
