+++ backend/src/chatroom/chatroom.gateway.ts
@@ -1,0 +1,6 @@
+import { SubscribeMessage, WebSocketGateway } from '@nestjs/websockets';
+import { Room } from './chatroom.entity';
+export let rooms = new Map<number, Room>();
+
+@WebSocketGateway()
+export class ChatroomGateway {}
