+++ backend/src/chatroom/chatroom.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { Room } from './chatroom.entity';
+import { ChatroomGateway } from './chatroom.gateway';
+
+
+@Module({
+  providers: [ChatroomGateway]
+})
+export class ChatroomModule {}
